backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai-whisper==20231117
faster-whisper==0.10.0
srt==3.5.3
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
ffmpeg-python==0.2.0
